refactor: move image cropper components view model , interaction listener... to component package
