+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/imageCropper/ImageCropperScreen.kt
@@ -25,7 +25,7 @@ import net.thechance.mena.designsystem.presentation.component.appBar.AppBar
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
 import net.thechance.mena.identity.presentation.base.BaseScreen
 import net.thechance.mena.identity.presentation.components.snackBar.IdentitySnackBarController
-import net.thechance.mena.identity.presentation.screen.imageCropper.components.ImageCropperComponent
+import net.thechance.mena.identity.presentation.screen.imageCropper.components.imageCropperComponent.ImageCropperComponent
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
 import org.koin.core.parameter.parametersOf
