+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/imageCropper/components/imageCropperComponent/ImageCropperComponentEffect.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.identity.presentation.screen.imageCropper
+package net.thechance.mena.identity.presentation.screen.imageCropper.components.imageCropperComponent
 
 sealed interface ImageCropperComponentEffect {
     data class SaveImage(val imageByteArray: ByteArray) : ImageCropperComponentEffect
